correct and test send_function

+++ users/models.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import AbstractUser
 class User(AbstractUser):
     username = None
     email = models.EmailField(unique=True, verbose_name='email')
-    tg_chat_id = models.IntegerField(verbose_name="Chat_id пользователя", null=True, blank=True)
+    tg_chat_id = models.BigIntegerField(verbose_name="Chat_id пользователя", null=True, blank=True)
     tg_username = models.CharField(max_length=50, verbose_name="телеграм юзернейм")
 
     USERNAME_FIELD = 'email'
